JSDT+BOWER Update Site
Includes the following 'bower'-related plug-ins:

- org.eclipse.wst.jsdt.nodejs.core
- org.eclipse.wst.jsdt.bower.core
- org.eclipse.wst.jsdt.bower.ide
- org.eclipse.wst.jsdt.bower.ide.ui

as well as according source plug-ins.

The local repository is added as well as Zip-archived site.

When installing (there will be a conflict) use 'Update my installation to be compatible with the items being installed' option.
This will install new versions of JavaScript Development Tools and SDK.

Signed-off-by: vrubezhny <vrubezhny@exadel.com>